refactor: move bridgeProviderAtom to entities
