+++ apps/cowswap-frontend/src/entities/bridgeProvider/index.ts
@@ -1,0 +1,2 @@
+export { useBridgeSupportedNetworks } from './useBridgeSupportedNetworks'
+export { useBridgeSupportedTokens } from './useBridgeSupportedTokens'
