+++ apps/cowswap-frontend/src/modules/bridge/hooks/index.ts
@@ -1,6 +1,3 @@
-export { useBridgeProvider } from './useBridgeProvider'
-export { useBridgeSupportedTokens } from './useBridgeSupportedTokens'
-export { useBridgeSupportedNetworks } from './useBridgeSupportedNetworks'
 export { useShouldDisplayBridgeDetails } from './useShouldDisplayBridgeDetails'
 export { useQuoteSwapContext } from './useQuoteSwapContext'
 export { useQuoteBridgeContext } from './useQuoteBridgeContext'
