+++ apps/cowswap-frontend/src/modules/twap/services/createTwapOrderTxs.test.ts
@@ -43,6 +43,7 @@ describe('Create TWAP order', () => {
     approveFn = jest.fn().mockReturnValue(APPROVE_TX_DATA)
 
     context = {
+      chainId,
       composableCowContract: {
         interface: { encodeFunctionData: createCowFn },
         address: COMPOSABLE_COW_ADDRESS[chainId],
